refactor: 인삿말 성공/실패 타입 도메인에 맞게 이름 수정 (#450)
