+++ module-independent/src/main/java/com/depromeet/type/greeting/GreetingSuccessType.java
@@ -1,14 +1,14 @@
-package com.depromeet.type.ai;
+package com.depromeet.type.greeting;
 
 import com.depromeet.type.SuccessType;
 
-public enum AISuccessType implements SuccessType {
+public enum GreetingSuccessType implements SuccessType {
     GET_RESPONSE_SUCCESS("AI_1", "AI 응답 받기에 성공하였습니다");
 
     private final String code;
     private final String message;
 
-    AISuccessType(String code, String message) {
+    GreetingSuccessType(String code, String message) {
         this.code = code;
         this.message = message;
     }
